All Registration Check Functions
All of the initial registration error check functions modified and condensed into a single file. Most have been modified to remove print screen checks that were useful during testing, but are not necessary for implementation. The data from the Registration form's entry boxes will be sent to each of the checks individually, and will either pass back an error, or check for duplicates and pass the info on to a .csv file for storage prior to sending the .csv file to the database.